fix(recipe): Change nbPerson type from Integer to Short for better memory efficiency

+++ api/src/main/java/com/a_table/model/RecipeEntity.java
@@ -26,7 +26,7 @@ public class RecipeEntity {
 
     private String name;
 
-    private Integer nbPerson;
+    private Short nbPerson;
 
     private String category;
